app/ai_models/vision.py
import torch
import torchvision
from PIL import Image
import numpy as np
from transformers import CLIPProcessor, CLIPModel
from ultralytics import YOLO

class VisionAnalyzer:
    def __init__(self):
        self.clip_model = CLIPModel.from_pretrained("openai/clip-vit-base-patch32")
        self.clip_processor = CLIPProcessor.from_pretrained("openai/clip-vit-base-patch32")
        self.yolo_model = YOLO('yolov8n.pt')
        
    def analyze_image(self, image_path):
        """
        Analyze medical images for abnormalities and conditions
        """
        try:
            # Load and preprocess image
            image = Image.open(image_path)
            
            # Run CLIP analysis for general medical conditions
            inputs = self.clip_processor(images=image, return_tensors="pt")
            image_features = self.clip_model.get_image_features(**inputs)
            
            # Run YOLO for object detection and segmentation
            results = self.yolo_model(image)
            
            # Combine results
            analysis = {
                "features": image_features.detach().numpy().tolist(),
                "detections": results[0].boxes.data.tolist(),
                "segmentation": results[0].masks.data.tolist() if results[0].masks else None
            }
            
            return analysis
            
        except Exception as e:
            return {"error": str(e)}
            
    def detect_skin_conditions(self, image_path):
        """
        Specialized analysis for skin conditions
        """
        try:
            image = Image.open(image_path)
            # Add specialized skin condition detection logic
            return {"status": "success", "conditions": []}
        except Exception as e:
            return {"error": str(e)}
            
    def analyze_medical_imaging(self, image_path, modality="xray"):
        """
        Analyze medical imaging (X-ray, MRI, CT)
        """
        try:
            image = Image.open(image_path)
            # Add specialized medical imaging analysis logic
            return {"status": "success", "findings": []}
        except Exception as e:
            return {"error": str(e)} 