requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
transformers==4.35.2
Pillow==11.2.1
numpy==1.26.2
opencv-python==4.8.1.78
ultralytics==8.0.196
scipy==1.16.0rc2
python-jose==3.3.0
passlib==1.7.4
python-dotenv==1.0.0 